backend/requirements.txt
# Ocean Sentinel Backend Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
supabase==1.3.0
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Clients & APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI & ML
google-generativeai==0.3.1
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4

# Blockchain Integration
web3==6.12.0
eth-account==0.9.0

# Real-time & Messaging
pusher==3.3.2
twilio==8.10.1
resend==0.6.0

# Environmental Data Processing
python-dateutil==2.8.2
pytz==2023.3
geopy==2.4.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Production WSGI
gunicorn==21.2.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Processing
Pillow==10.1.0
python-magic==0.4.27

# Cryptography
cryptography==41.0.8
hashlib-compat==1.0.1

# Email Templates
jinja2==3.1.2

# Geographic Data
Shapely==2.0.2
geojson==3.1.0

# API Documentation
fastapi-users==12.1.2

# Environment Variables
python-decouple==3.8

# Time Series Data
arrow==1.3.0

# Data Export
xlsxwriter==3.1.9
openpyxl==3.1.2

# Cache
redis==5.0.1
python-memcached==1.59

# Health Checks
fastapi-health==0.4.0

# Metrics
prometheus-client==0.19.0
